+++ include/state-observation/dynamical-system/dynamical-system-functor-base.hpp
@@ -48,11 +48,11 @@ namespace stateObservation
         virtual void reset(){}
 
         ///gets the state size
-        virtual unsigned getStateSize()const =0 ;
+        virtual size_t getStateSize()const =0 ;
         ///gets the input size
-        virtual unsigned getInputSize()const =0 ;
+        virtual size_t getInputSize()const =0 ;
         ///gets the measurements size
-        virtual unsigned getMeasurementSize() const=0;
+        virtual size_t getMeasurementSize() const=0;
 
         ///Gives a boolean answer on whether or not the vector is correctly sized to be a state vector
         virtual bool checkStateVector(const Vector &);
