+++ include/state-observation/observer/zero-delay-observer.hpp
@@ -50,7 +50,7 @@ namespace stateObservation
         ///  \li n : size of the state vector
         ///  \li m : size of the measurements vector
         ///  \li p : size of the input vector
-        ZeroDelayObserver(unsigned n, unsigned m, unsigned p=0):
+        ZeroDelayObserver(size_t n, size_t m, size_t p=0):
             ObserverBase(n,m,p){}
 
         ///Default constructor (default values for n,m,p are zero)
@@ -125,13 +125,13 @@ namespace stateObservation
         virtual TimeSize getMeasurementsNumber()const;
 
         ///changes the size of the state vector: resets the stored state vector
-        virtual void setStateSize(unsigned n);
+        virtual void setStateSize(size_t n);
 
         ///changes the size of the measurement vector: reset the stored measurement vectors
-        virtual void setMeasureSize(unsigned m);
+        virtual void setMeasureSize(size_t m);
 
         ///changes the size of the input vector: reset the stored input vectors
-        virtual void setInputSize(unsigned p);
+        virtual void setInputSize(size_t p);
 
     protected:
