+++ include/state-observation/tools/definitions.hpp
@@ -564,9 +564,9 @@ namespace detail
 
     std::string STATE_OBSERVATION_DLLAPI vectorToString(const Vector& v);
 
-    Matrix STATE_OBSERVATION_DLLAPI stringToMatrix(const std::string& str, unsigned rows, unsigned cols);
+    Matrix STATE_OBSERVATION_DLLAPI stringToMatrix(const std::string& str, size_t rows, size_t cols);
 
-    Vector STATE_OBSERVATION_DLLAPI stringToVector(const std::string& str, unsigned length);
+    Vector STATE_OBSERVATION_DLLAPI stringToVector(const std::string& str, size_t length);
 
     Vector STATE_OBSERVATION_DLLAPI stringToVector(const std::string& str);
   }
